requirements.txt
agora-token-builder==1.0.0
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
crispy-bootstrap5==0.7
Django==4.0.4
django-bootstrap5==22.2
django-admin-interface==0.19.1
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-colorfield==0.7.0
django-cors-headers==3.13.0
django-dirtyfields==1.9.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-smart-selects==1.6.0
django-tables2==2.4.1
djangorestframework==3.13.1
firebase-admin==6.0.1
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.21.0
idna==3.4
msgpack==1.0.4
mysqlclient==2.1.0
Pillow==9.3.0
proto-plus==1.22.1
protobuf==4.21.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.1
pytz==2022.6
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
